server/requirements.txt
asgiref==3.7.2
coverage==7.3.2
dj-database-url==2.1.0
Django==4.1.4
django-cors-headers==4.3.0
django-heroku==0.3.1
djangorestframework==3.12.4
flake8==6.1.0
gunicorn==21.2.0
mccabe==0.7.0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.5
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2021.3
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
